acquireAuditData {rattle}	R Documentation

Generate the audit dataset.

Description

Rattle uses an artificial dataset for demonstration purposes. This function retrieves the source data ftp://ftp.ics.uci.edu/pub/machine-learning-databases/adult/adult.data and then transforms the data in a variety of ways.

Usage

acquireAuditData(write.to.file=FALSE)

Arguments

write.to.file
Whether to generate a colleciton of files based on the data. The files generated include: audit.csv, audit.Rdata, audit.arf, and audit\_missing.csv

Details

See the function definition for details of the processing done on the data downloaded from the UCI repository.

Value

By default the function returns a data frame containing the audit dataset. If write.to.file is TRUE then the data frame is returned invisibly.

Author(s)

Graham.Williams@togaware.com

References

Package home page: http://rattle.togaware.com

See Also

audit, rattle.

[Package rattle version 5.0.14 Index]