asRules.rpart {rattle}	R Documentation

List the rules corresponding to the rpart decision tree

Description

Display a list of rules for an rpart decision tree.

Usage

## S3 method for class 'rpart'
asRules(model, compact=FALSE, ...)

Arguments

model
an rpart model.

compact
whether to list cateogricals compactly.

...
further arguments passed to or from other methods.

Details

Traverse a decision tree to generate the equivalent set of rules, one rule for each path from the root node to a leaf node.

Author(s)

Graham.Williams@togaware.com

References

Package home page: http://rattle.togaware.com

Examples

## Not run: asRules.rpart(my.rpart)

[Package rattle version 5.0.14 Index]