binning {rattle}	R Documentation

Perform binning over numeric data

Description

Perform binning.

Usage

binning(x, bins=4, method=c("quantile", "wtd.quantile", "kmeans"),
                     labels=NULL, ordered=TRUE, weights=NULL)

Arguments

x
the numeric data to bin.

bins
the number of bins to use.

method
whether to use "quantile", weighted quantile "wtd.quantile" or "kmeans" binning.

labels
the labels or names to use for each of the bins.

ordered
whether to build an ordered factor or not.

weights
vector of numeric weights for each observation for weighted quantile binning.

Details

Bin the provided nmeric data into the specified number of bins using one of the supported methods. The bins will have the names specified by labels, if supplied. The result can optionally be an ordered factor.

Value

A factor is returned.

Author(s)

Daniele Medri and Graham Williams

References

Package home page: http://rattle.togaware.com

[Package rattle version 5.0.14 Index]