calcInitialDigitDistr {rattle}	R Documentation

Generate a frequency count of the initial digits

Description

In the context of Benford's Law calculate the distribution of the frequencies of the first digit of the numbers supplied as the argument.

Usage

calcInitialDigitDistr(l, digit=1, len=1,
 sp=c("none", "positive", "negative"))

Arguments

l
a vector of numbers.

digit
the digit to generate frequencies for.

len
The number of digits.

sp
whether and how to split the digits.

Author(s)

Graham.Williams@togaware.com

References

Package home page: http://rattle.togaware.com

[Package rattle version 5.0.14 Index]