centers.hclust {rattle}	R Documentation

List Cluster Centers for a Hierarchical Cluster

Description

Generate a matrix of centers from a hierarchical cluster.

Usage

centers.hclust(x, h, nclust=10, use.median=FALSE)

Arguments

x
The data used to build the cluster.

h
A hclust object.

nclust
Number of clusters.

use.median
Use meadion instead of mean.

Details

For the specified number of clusters, cut the hierarchical cluster appropriately to that number of clusters, and return the mean (or median) of each resulting cluster.

Author(s)

Daniele Medri and Graham.Williams@togaware.com

References

Package home page: http://rattle.togaware.com

[Package rattle version 5.0.14 Index]