drawTreeNodes {rattle}	R Documentation

Draw nodes of a decision tree

Description

Draw the nodes of a decision tree

Usage

drawTreeNodes(tree, cex = par("cex"), pch = par("pch"),
                           size = 4 * cex, col = NULL, nodeinfo = FALSE,
                           units = "", cases = "obs", 
                           digits = getOption("digits"),
                           decimals = 2,
                           print.levels = TRUE, new = TRUE) 

Arguments

tree
an rpart decision tree.

cex
.

pch
.

size
.

col
.

nodeinfo
.

units
.

cases
.

digits
.

decimals
the number of decimal digits to include in numeric split nodes.

print.levels
.

new
.

Details

A variation of draw.tree() from the maptree package.

Author(s)

Graham.Williams@togaware.com, Denis White

References

Package home page: http://rattle.togaware.com

Examples

## this is usually used in the context of the plotRisk function
## Not run: drawTreeNodes(rpart(Species ~ ., iris))

[Package rattle version 5.0.14 Index]