listTreesAda {rattle}	R Documentation

List trees from an Ada model

Description

Display the textual representation of a selection of Ada trees.

Usage

listTreesAda(model, trees=0)

Arguments

model
an ada model.

trees
The list of trees to list. Use 0 to list all trees.

Details

Using rpart's print method display each of the specified trees from an Ada model.

Author(s)

Graham.Williams@togaware.com

References

Package home page: http://rattle.togaware.com

Examples

## Not run: listTreesAda(ds.ada)

[Package rattle version 5.0.14 Index]