listVersions {rattle}	R Documentation

Versions of Installed Packages

Description

Generate a list of packages installed and their version number.

Usage

listVersions(file="", ...)

Arguments

file
a character string naming a file or a connection open for writing. '""' indicates output to the console.

...
arguments to write.csv.

Details

This function is useful in reporting problems or bugs, to ensure there is a clear match of R package versions between the system exhibiting the issue and the test system replicating the issue.

By default the information is written to the console in a comma separated form, that is ideally designed to be written to a CSV file for emailing.

Author(s)

Graham.Williams@togaware.com

See Also

write.csv

[Package rattle version 5.0.14 Index]