randomForest2Rules {rattle}	R Documentation

Generate accessible data structure of a randomForest model

Description

A randomForest model, by default, consists of 500 decision trees. This function walks through each tree and generates a set of rules. This takes a considerable amount of time and is provided for users to access the actual model, but it is not yet used within the Rattle GUI. It may be used to display the output of the RF (but it takes longer to generate than the model itself!). Or it might only be used on export to PMML or SQL.

Usage

randomForest2Rules(model, models=NULL)

Arguments

model
a randomForest model.

models
a list of integers limiting the models in MODEL that are converted.

Author(s)

Graham.Williams@togaware.com

References

Package home page: http://rattle.togaware.com

Examples

## Generate a ruleset for a specific model amongst the 500.
## Not run: randomForest2Rules(rfmodel, 5)

## Generate a ruleset for specific models amongst the 500.
## Not run: randomForest2Rules(rfmodel, c(5,10,15))

## Generate a ruleset for each of the 500 models.
## Not run: randomForest2Rules(rfmodel)

[Package rattle version 5.0.14 Index]