savePlotToFile {rattle}	R Documentation

Save a plot in some way

Description

For the current device, or for the device identified, save the plot displayed there in some way. This is either saved to file, copied to the clipboard for pasting into other applications, or sent to the printer for saving a hard copy.

Usage

savePlotToFile(file.name, dev.num=dev.cur())
copyPlotToClipboard(dev.num=dev.cur())
printPlot(dev.num=dev.cur()) 

Arguments

file.name
Character string naming the file including the file name extension which is used to specify the type of file to save.

dev.num
A device number indicating which device to save.

Author(s)

Graham.Williams@togaware.com

References

Package home page: http://rattle.togaware.com

[Package rattle version 5.0.14 Index]