setupDataset {rattle}	R Documentation

Given specific contents of env add other dataset related variables.

Description

This rattle support function is used for encapsulating data mining objects. The supplied environment is augmented with other data derived from the supplied data, such as a sample trianing dataset, list of numeric variables, and a formula for modelling.

Usage

setupDataset(env, seed=NULL)

Arguments

env
the environment to modify.

seed
optionally set the seed for repeatability.

Details

The supplied object (an environment) is assumed to also contain the variables data (a data frame), target (a character string naming the target variable), risk (a character string naming the risk variable), and inputs (a character vector naming all the input variables). This function then adds in the variables vars (the variables used for modelling), numerics (the numeric vars within inputs), nobs (the number of observations), form (the formula for building models), train (a 70% training dataset).

Author(s)

Graham.Williams@togaware.com

References

Package home page: http://rattle.togaware.com

[Package rattle version 5.0.14 Index]