treeset.randomForest {rattle}	R Documentation

Generate a representation of a tree in a Random Forest

Description

Often we want to view the actual trees built by a random forest. Although reviewing all 500 trees might be a bit much, this function allows us to at least list them.

Usage

treeset.randomForest(model, n=1, root=1, format="R")

Arguments

model
a randomForest model.

n
a specific tree to list.

root
where to start the stree from, primarily for internal use.

format
one of "R", "VB".

Author(s)

Graham.Williams@togaware.com

References

Package home page: http://rattle.togaware.com

Examples

## Display a treeset for a specific model amongst the 500.
## Not run: treeset.randomForests(rfmodel, 5)

[Package rattle version 5.0.14 Index]