weather {rattle}	R Documentation

Sample and full dataset for illustrating Rattle functionality.

Description

Weather observations from a number of locations around Australia, obtained from the Australian Commonwealth Bureau of Meteorology and processed to create a sample dataset for illustrating data mining using R and Rattle.

The data has been processed to provide a target variable RainTomorrow (whether there is rain on the following day - No/Yes) and a risk variable RISK_MM (how much rain). Various transformations are performed on the data.

The weather dataset is a fixed subset of the full weatherAUS dataset, for the purpose of repeatable demonstration. It includes only one year of data for a single weather station.

The weatherAUS dataset is ever changing, with each release of Rattle. It is updated from the Bureau of Meteorology web site regularly.

The locationsAUS dataset records the location of each weather station.

The source dataset is Copyright by the Australian Commonwealth Bureau of Meteorology and used with permission.

Usage

weather
weatherAUS

Format

The weather dataset is a data frame containing one year of daily observations from a single weather station (Canberra). The weatherAUS dataset is a data frame containing over 35,000 daily observations from over 45 Australian weather stations.

Date

The date of observation (a Date object).

Location

The common name of the location of the weather station.

MinTemp

The minimum temperature in degrees celsius.

MaxTemp

The maximum temperature in degrees celsius.

Rainfall

The amount of rainfall recorded for the day in mm.

Evaporation

The so-called Class A pan evaporation (mm) in the 24 hours to 9am.

Sunshine

The number of hours of bright sunshine in the day.

WindGustDir

The direction of the strongest wind gust in the 24 hours to midnight.

WindGustSpeed

The speed (km/h) of the strongest wind gust in the 24 hours to midnight.

Temp9am

Temperature (degrees C) at 9am.

RelHumid9am

Relative humidity (percent) at 9am.

Cloud9am

Fraction of sky obscured by cloud at 9am. This is measured in "oktas", which are a unit of eigths. It records how many eigths of the sky are obscured by cloud. A 0 measure indicates completely clear sky whilst an 8 indicates that it is completely overcast.

WindSpeed9am

Wind speed (km/hr) averaged over 10 minutes prior to 9am.

Pressure9am

Atmospheric pressure (hpa) reduced to mean sea level at 9am.

Temp3pm

Temperature (degrees C) at 3pm.

RelHumid3pm

Relative humidity (percent) at 3pm.

Cloud3pm

Fraction of sky obscured by cloud (in "oktas": eighths) at 3pm. See Cload9am for a description of the values.

WindSpeed3pm

Wind speed (km/hr) averaged over 10 minutes prior to 3pm.

Pressure3pm

Atmospheric pressure (hpa) reduced to mean sea level at 3pm.

ChangeTemp

Change in temperature.

ChangeTempDir

Direction of change in temperature.

ChangeTempMag

Magnitude of change in temperature.

ChangeWindDirect

Direction of wind change.

MaxWindPeriod

Period of maximum wind.

RainToday

Integer: 1 if precipitation (mm) in the 24 hours to 9am exceeds 1mm, otherwise 0.

TempRange

Difference between minimum and maximum temperatures (degrees C) in the 24 hours to 9am.

PressureChange

Change in pressure.

RISK_MM

The amount of rain. A kind of measure of the "risk".

RainTomorrow

The target variable. Did it rain tomorrow?

Author(s)

Graham.Williams@togaware.com

Source

Observations were drawn from numerous weather stations. The daily observations are available from http://www.bom.gov.au/climate/data. Copyright Commonwealth of Australia 2010, Bureau of Meteorology.

Definitions adapted from http://www.bom.gov.au/climate/dwo/IDCJDW0000.shtml

References

Package home page: http://rattle.togaware.com. Data source: http://www.bom.gov.au/climate/dwo/ and http://www.bom.gov.au/climate/data.

See Also

audit.

[Package rattle version 5.0.14 Index]